Add result type for use in completion handler
Provides bit safer static guarantees about the result of a request. Right now it's just being used for network requests, but is generic enough to be used for any failable operations. This gives compile time checking in the `send` function that a valid completion handler will be called (with either a `response` object or an `error` object). Right now this isn't much of an issue since the function isn't that complicated, but if more control paths pop up in the future this guarantees that either a success or fail is passed back to the caller. 

It also statically forces the caller to handle both cases, success or failure, which is good practice. Even if handling an error is just to ignore it, it's best to explicitly ignore it so in the future it can be known as a conscious decision.